refactor code Savko
