+++ src/app/styles/index.scss
@@ -1,3 +1,3 @@
-@forward 'colors';
+@forward 'variables';
 @forward 'typography';
 @forward 'boilerplate';
