+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import '@/styles/index.scss'
+import '@/app/styles/index.scss'
 import type { AppProps } from 'next/app'
 
 export default function App({ Component, pageProps }: AppProps) {
