+++ src/shared/assets/icons/check.tsx
