+++ src/shared/assets/icons/eye-closed.tsx
