+++ src/shared/assets/icons/eye.tsx
