+++ src/shared/assets/icons/icon-wrapper.tsx
