+++ src/shared/assets/icons/keyboard-arrow-left.tsx
