+++ src/shared/assets/icons/keyboard-arrow-right.tsx
