+++ src/shared/assets/icons/log-out-icon.tsx
