+++ src/shared/assets/icons/search-icon.tsx
