+++ src/shared/ui/checkbox/checkbox.module.scss
