+++ src/shared/ui/checkbox/index.ts
@@ -1,0 +1,1 @@
+export { CheckboxItem } from './checkbox'
