+++ src/shared/ui/input/input.stories.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, useState } from 'react'
 
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { Input } from './index'
+import { Input } from './input'
 
 const meta = {
   title: 'Components/Input',
