+++ src/shared/ui/select/index.ts
@@ -1,0 +1,1 @@
+export { Select } from './select'
