+++ src/shared/ui/typography/typography.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react'
 
-import { Typography } from './index'
+import { Typography } from './typography'
 
 const meta = {
   title: 'Components/Typography1',
