--- src/ui/components/input/index.ts
@@ -1,1 +1,0 @@
-export * from './input'
