--- src/ui/components/select/index.ts
@@ -1,1 +1,0 @@
-export * from './select'
