--- src/ui/components/typography/index.ts
@@ -1,1 +1,0 @@
-export * from './typography'
